Add files via upload
Added error2.m, that gives the percentCorrect, onesPercentCorrect, total ones (in yTest), and onesGuessed (number of 1s in yPTest).

Updated kaggledata.m to do some classifiers, like svm, decision trees, treebagger, adaboost. If any of these are done incorrectly, please tell me what I need to correct. Also added XTest and stuff of the first 80% of the data so we can see how the percentages of score is when we do 'ScoreTransform' to our classifiers. This way we know where the classifiers guesses 1 and we can take the ScoreTransform percentages that fit best.